Fixes assorted bugs with the thread tracker. Fixes a few bugs in the SimMap connectors. Refactors part of the base of aurora.

+++ Aurora/Modules/World/Map/RemoteSimMapConnector.cs
@@ -282,7 +282,7 @@ namespace Aurora.Modules
             Dictionary<string, object> sendData = new Dictionary<string, object>();
 
             sendData["REGIONHANDLE"] = regionHandle;
-            sendData["GRIDITEMTYPE"] = gridItemType;
+            sendData["GRIDITEMTYPE"] = (int)gridItemType;
             sendData["METHOD"] = "getmapitems";
 
             string reqString = ServerUtils.BuildQueryString(sendData);
